Refactor pin_creator and add HasOneMatcher + HasAndBelongsToManyMatcher
Logic found in Ruby code is now packaged as a Matcher class under /meta_source.

MetaSource::Associations::BelongsToMatcher, for instance, is responsible for
identifying belongs_to lines and extracting the name and type of the reader method.

Currently it's assumed that each Matcher will only be responsible for one pin,
but this will change when more association support is added, e.g. there will
need to be pins created for the reader method and the setter method (plus
others) when belongs_to is encountered.