Do not exit after bundle install / yard gems error
Yard gems exits with `1` in ruby 3.1.0 builds; since that doesn't seem to cause
the tests to fail, just continue the build.

This is a cheeky option, we should probably split the bundle install and yard
gems commands, or do a check for the version of Ruby. My bash skills were
failing me so now I am trying this.

The error (visible when running with --debug) looks like:

```
~/.asdf/installs/ruby/3.1.0/lib/ruby/gems/3.1.0/gems/yard-0.9.28/lib/yard/parser/ruby/ast_node.rb:274:in `first': cannot get the first element of beginless range (RangeError)
        from /Users/frederickmeissner/.asdf/installs/ruby/3.1.0/lib/ruby/gems/3.1.0/gems/yard-0.9.28/lib/yard/parser/ruby/ast_node.rb:274:in `line'
        from /Users/frederickmeissner/.asdf/installs/ruby/3.1.0/lib/ruby/gems/3.1.0/gems/yard-0.9.28/lib/yard/parser/ruby/ruby_parser.rb:624:in `block in insert_comments'
        from /Users/frederickmeissner/.asdf/installs/ruby/3.1.0/lib/ruby/gems/3.1.0/gems/yard-0.9.28/lib/yard/parser/ruby/ast_node.rb:212:in `traverse'
        from /Users/frederickmeissner/.asdf/installs/ruby/3.1.0/lib/ruby/gems/3.1.0/gems/yard-0.9.28/lib/yard/parser/ruby/ruby_parser.rb:615:in `insert_comments'
        ...
 ```

Which seems a lot like this issue in the YARD project:
https://github.com/lsegal/yard/issues/1327

However that is ostensibly fixed in yard v0.9.25 and I am running v0.9.28. I'm
not interested in debugging YARD and this doesn't feel like an awful workaround.