Clean up spec/definitions/**/*.yml
These contained a load of duplicated definitions for methods that are
inherited from Module. This will make tests a teensy tiny bit faster,
but mostly makes these files actually sort of useful.

These changes were made manually, the generate_definitions.rb script
should probably be updated to be a bit more precise, but we can deal
with that the next time somebody feels like running it.